% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openController}
\alias{openController}
\title{Open Keyboard (in the viewer pane if in RStudio).}
\usage{
openController(host = getOption("plotAR.external.url"))
}
\description{
Open Keyboard (in the viewer pane if in RStudio).
}
