% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar.R
\name{plotAR}
\alias{plotAR}
\title{plotAR}
\usage{
plotAR(
  data,
  x,
  y,
  z,
  col = NULL,
  size = NULL,
  type = "p",
  lines = NULL,
  label = NULL,
  name = NULL,
  description = NULL,
  speed = 0L,
  autoScale = T,
  digits = 5,
  doOpenController = TRUE,
  .send = TRUE
)
}
\description{
plotAR
}
