% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar.R
\name{plotAR_}
\alias{plotAR_}
\title{Plot data into the Cardboard VR}
\usage{
plotAR_(
  data,
  x = NULL,
  y = NULL,
  z = NULL,
  col = NULL,
  size = NULL,
  label = NULL,
  name = NULL,
  description = NULL,
  axis_names = NULL,
  col_labels = NULL,
  speed = 0L,
  autoScale = TRUE,
  type = "p",
  lines = NULL,
  digits = 5,
  doOpenController = TRUE,
  .send = TRUE
)
}
\arguments{
\item{data}{data frame of which the first three columns will be used in the scatter.}

\item{col}{the color}

\item{size}{the size}

\item{label}{text labels of data}

\item{name}{the name of this dataset - by default will be inferred from \code{data}.}

\item{description}{description of the data}

\item{speed}{the speed along which the viewer will fly through data}

\item{autoScale}{standardize data before sending}

\item{type}{point (p) or line (l)}

\item{lines}{if you want to have lines addition to data}

\item{digits}{precision of data sent}

\item{doOpenController}{after sending, should the controller be (re-)opened?}

\item{.send}{send the data}
}
\description{
This is the main function for usage of plotAR.
If necessary it starts a deamonized server and posts the data on that server.
If no controller is open one also will be opened.
All connected devices will be informed of the new data and should reload the data.
}
\examples{
\dontrun{
plotAR(iris, col=iris$Species)
}
}
\seealso{
\code{\link{open}}
}
