% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRdboard.R
\name{startBlockingServer}
\alias{startBlockingServer}
\title{Start the WebServer in a blocking version.}
\usage{
startBlockingServer(
  host = "0.0.0.0",
  port = 2908,
  interruptIntervalMs = ifelse(interactive(), 100, 1000)
)
}
\arguments{
\item{host}{on which host to listen}

\item{port}{on which port to listen}

\item{interruptIntervalMs}{passed to \code{httpuv}}
}
\description{
This blocks until interrupted or killed. This is the actual server
that gets run by \code{\link{startServer}} in a background process.
}
