% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotar.R
\name{startServer}
\alias{startServer}
\title{Start Server}
\usage{
startServer(port = 2908L)
}
\arguments{
\item{...}{passed to \code{callr::r_bg}}
}
\value{
the process object (invisibly)
}
\description{
Start Server
}
